Widen background job duration columns to bigint (#929)
"Clear stranded" failed in production with a rolled-back transaction. The
trace shows the job cancel itself succeeding -- BEGIN, SELECT 2000 targets,
UPDATE 2000 jobs, recalc SELECT -- and then the BackgroundBatch summary
UPDATE throwing 22003:

  value "5184000000" is out of range for type integer
  unnamed portal parameter $14   ("wallDurationMs")

int32 caps these columns at 2_147_483_647 ms = 24.86 days, and they hold
wall-clock spans over a batch's entire life. wallDurationMs is
firstEnqueuedAt -> lastJobFinishedAt, so cancelling jobs enqueued 60 days
ago stamped finishedAt = now and produced 5.18e9. Targets are ordered
createdAt ASC, so the oldest stranded batch (2026-06-28) is recalculated
first and takes the whole transaction down with it.

This is a latent bug in recalculateBackgroundBatchSummaryTx rather than
anything specific to the new cleanup path -- cancelQueuedJobsForBatch has
the same exposure, and any batch left open longer than 24.86 days hits it
the moment a job of its gets a finishedAt. Losing a queue wakeup is exactly
what leaves a batch open that long, so the durability bug and this one
share a cause.

- BackgroundBatch.processingDurationMs / wallDurationMs / activeDurationMs
  and BackgroundJob.durationMs are now bigint.
- The SQL casts producing them widen to ::bigint. BackgroundJob.durationMs
  is GREATEST(0, (now - startedAt) * 1000), which overflows the same way
  for a job whose worker died and held `running` past the lease -- there
  are 17 such rows in production right now, so the sweeper reclaiming them
  would have hit this next.

Count casts stay ::int; those are row counts, not durations.

Requires the column widening on production before this deploys. Widening is
backward compatible in that direction: currently-deployed code writes int
values that fit a bigint column unchanged.

+++ apps/web/src/server/repo/background-jobs.ts
@@ -694,7 +694,7 @@ export async function finishBackgroundJob(
       .set({
         status,
         finishedAt: now,
-        durationMs: sql`GREATEST(0, FLOOR(EXTRACT(EPOCH FROM (${now}::timestamp - ${backgroundJob.startedAt})) * 1000))::int`,
+        durationMs: sql`GREATEST(0, FLOOR(EXTRACT(EPOCH FROM (${now}::timestamp - ${backgroundJob.startedAt})) * 1000))::bigint`,
       })
       .where(eq(backgroundJob.id, jobId))
       .returning({ batchId: backgroundJob.batchId });
@@ -728,7 +728,7 @@ export async function failOrRetryBackgroundJob(
         finishedAt: nextStatus === "failed" ? now : null,
         durationMs:
           nextStatus === "failed"
-            ? sql`GREATEST(0, FLOOR(EXTRACT(EPOCH FROM (${now}::timestamp - ${backgroundJob.startedAt})) * 1000))::int`
+            ? sql`GREATEST(0, FLOOR(EXTRACT(EPOCH FROM (${now}::timestamp - ${backgroundJob.startedAt})) * 1000))::bigint`
             : null,
         lastError: getErrorMessage(error),
       })
@@ -957,7 +957,7 @@ async function recalculateBackgroundBatchSummaryTx(
       MAX("queuedAt") AS "lastEnqueuedAt",
       MIN("startedAt") AS "firstJobStartedAt",
       MAX("finishedAt") AS "lastJobFinishedAt",
-      COALESCE(SUM("durationMs"), 0)::int AS "activeDurationMs"
+      COALESCE(SUM("durationMs"), 0)::bigint AS "activeDurationMs"
     FROM "BackgroundJob"
     WHERE "batchId" = ${batchId}
       AND "deletedAt" IS NULL
